requirements.txt
cryptography
Pillow
piexif
